Fixed Page<T> extensions.
In case the client doesn't set the Paging property, WCF will deserialize that property as null, even though the property was initialized with an instance in code. This only happens when WCF is used without the shared Contract assembly (and it doesn't use with JSON.NET deserializing). The Page<T> extensions will now handle the case where the Page<T> is a null reference.